feat(v1alpha1): tighten casting schema validation
Add validation keywords to the v1alpha1 schema and surface them at config
load time:

- Casting marks Metadata, Spec required and rejects unknown top-level
  fields via additionalProperties:false; CastingSpec marks Deployment
  required.
- TypeVersion.APIVersion and TypeMetadata.Name are required and
  non-nullable; APIVersion enum-pinned to "v1alpha1".
- MetaStoreKind implements jsonschema.Enum so kind validates against
  the registered MetaStoreKinds() list.
- Patch types and Merge() move to dedicated files (patch.go, merge.go);
  no behavior change.
- yamlconfig.GetV1Alpha1 now wraps validation failures with
  errors.TypeInvalidInput so callers can distinguish bad input from
  internal failures.
- Regenerated schema.json reflects the new constraints.
